chore: 패키지 구조 변경
- 전체 패키지 구조 webserver와 webapp으로 분리
- Controller 인터페이스, ControllerMapper, ViewResolver 제거
