--- src/main/java/controller/Controller.java
@@ -1,9 +1,0 @@
-package controller;
-
-import webserver.http.message.HttpRequest;
-import webserver.http.message.HttpResponse;
-
-public interface Controller {
-
-	public String process(HttpRequest request, HttpResponse response);
-}
