--- src/main/java/controller/HomeController.java
@@ -1,12 +1,0 @@
-package controller;
-
-import webserver.http.message.HttpRequest;
-import webserver.http.message.HttpResponse;
-
-public class HomeController implements Controller {
-
-	@Override
-	public String process(HttpRequest request, HttpResponse response) {
-		return "index";
-	}
-}
