+++ src/main/java/webapp/controller/HomeController.java
@@ -1,0 +1,14 @@
+package webapp.controller;
+
+import webserver.annotation.Controller;
+import webserver.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+
+	@GetMapping(path = "/index.html")
+	public String home() {
+		return "index";
+	}
+
+}
